requirements/common.txt
dj-database-url==0.4.1
Django==1.10
django-configurations==2.0
psycopg2==2.6.2
requests==2.11.1
whitenoise==3.2.1
